requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
flask-ngrok==0.0.25
frozenlist==1.3.3
future==0.18.3
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.3
MarkupSafe==2.1.3
multidict==6.0.4
pyimgur==0.6.0
pyngrok==6.0.0
PyYAML==6.0
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.15.0
